feat: add company bas role

+++ app/Http/Controllers/CompanyController.php
@@ -46,6 +46,7 @@ class CompanyController extends Controller
 
         // For all other roles: search all companies
         $companies = Company::query()
+            ->where('admin_id', $user->id)
             ->when($search, function ($query) use ($search) {
                 $query->where(function ($q) use ($search) {
                     $q->where('name', 'LIKE', '%' . $search . '%')
@@ -83,6 +84,7 @@ class CompanyController extends Controller
             'name' => $validated['company_name'],
             'email' => $validated['email'],
             'address' => $validated['address'],
+            'admin_id' => auth()->id()
         ]);
         return to_route('companies.list')->with('toast', 'Record add successfully');
         // return response()->json(['message' => 'Company created', 'company' => $company]);
@@ -90,14 +92,21 @@ class CompanyController extends Controller
 
     public function edit($id)
     {
+        $company = Company::findOrFail($id);
+        if ($company->admin_id != auth()->id()) {
+            return back()->with('toast', 'Cannot get other company');
+        }
         return Inertia::render('companies/Update', [
-            'company' => Company::findOrFail($id)
+            'company' => $company
         ]);
     }
 
     public function update(Request $request, $id)
     {
         $company = Company::findOrFail($id);
+        if ($company->admin_id != auth()->id()) {
+            return back()->with('toast', 'Cannot update other company');
+        }
         $validated = $request->validate([
             'company_name' => 'required|string|max:255',
             'email' => 'required|email|unique:companies,email,' . $company->id,
@@ -114,6 +123,9 @@ class CompanyController extends Controller
 
     public function destroy(Company $company)
     {
+        if ($company->admin_id != auth()->id()) {
+            return back()->with('toast', 'Cannot delete other record');
+        }
         $company->delete();
         return back()->with('toast', 'Record Delete successfully');
     }
